#224: Test files put into proper folder
