+++ Server/src/test/java/UnitTests/rul/models/AdditiveRegressionTest.java
@@ -1,22 +1,24 @@
-package rul.models;
+package UnitTests.rul.models;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import rul.models.AdditiveRegressionModelImpl;
+import rul.models.ModelsController;
 import weka.classifiers.Classifier;
-import weka.classifiers.functions.MultilayerPerceptron;
+import weka.classifiers.meta.AdditiveRegression;
 import weka.core.Instances;
 
 import java.io.FileReader;
 
 import static org.junit.Assert.assertEquals;
 
-public class MultilayerPerceptronTest {
+public class AdditiveRegressionTest {
     private ModelsController modelsController;
 
     @Before
     public void setUp() {
-        modelsController = new ModelsController(new MultilayerPerceptronModelImpl());
+        modelsController = new ModelsController(new AdditiveRegressionModelImpl());
     }
 
     @After
@@ -30,8 +32,8 @@ public class MultilayerPerceptronTest {
         Instances  trainData = new Instances(trainFile);
         trainData.setClassIndex(trainData.numAttributes() - 1);
 
-        Classifier mlp = new MultilayerPerceptron();
-        assertEquals("Should Return Multilayer Perceptron Model", mlp.getClass(),
+        Classifier additiveRegression = new AdditiveRegression();
+        assertEquals("Should Return Additive Regression Model", additiveRegression.getClass(),
                      modelsController.trainModel(trainData).getClass());
     }
 }
