+++ Server/src/test/java/UnitTests/rul/models/LinearRegressionTest.java
@@ -3,6 +3,8 @@ package UnitTests.rul.models;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import rul.models.LinearRegressionModelImpl;
+import rul.models.ModelsController;
 import weka.classifiers.functions.LinearRegression;
 import weka.core.Instances;
